Update `FileStorage` to save data in byte sequence format, not in JSON.
