+++ client/src/main/kotlin/io/spine/examples/pingh/client/FileStorage.kt
@@ -27,20 +27,19 @@
 package io.spine.examples.pingh.client
 
 import com.google.common.annotations.VisibleForTesting
-import com.google.gson.Gson
+import com.google.protobuf.Message
 import java.io.File
+import java.io.FileInputStream
+import java.io.FileOutputStream
 import kotlin.io.path.Path
 import kotlin.io.path.createDirectories
 import kotlin.io.path.createFile
 import kotlin.io.path.exists
-import net.harawata.appdirs.AppDirsFactory
 
 /**
- * A repository that stores data in JSON format on a file on disk.
+ * Stores data on disk in a sequence of bytes.
  */
 internal object FileStorage {
-    private val gson = Gson()
-
     /**
      * Returns the data from the storage file if it contains content;
      * otherwise, returns the `default`.
@@ -48,14 +47,21 @@ internal object FileStorage {
      * @param T The type of class being loaded.
      *
      * @param from The location of the file from which the data is read.
+     * @param parser Deserializes a byte sequence into a message.
      * @param default Calculates the default value to use when no data is found in the file.
      */
-    internal inline fun <reified T> loadOrDefault(from: FileLocation, default: () -> T): T {
-        val content = storage(from).readText()
-        return if (content.isNotBlank()) {
-            gson.fromJson(content, T::class.java)
-        } else {
-            default()
+    internal inline fun <reified T : Message> loadOrDefault(
+        from: FileLocation,
+        parser: (ByteArray) -> T,
+        default: () -> T
+    ): T {
+        FileInputStream(storage(from)).use { file ->
+            val bytes = file.readAllBytes()
+            return if (bytes.isNotEmpty()) {
+                parser(bytes)
+            } else {
+                default()
+            }
         }
     }
 
@@ -65,11 +71,12 @@ internal object FileStorage {
      * @param T The type of class being saved.
      *
      * @param to The location of the file to which the data is written.
-     * @param data The object that is converted to JSON and written to a file.
+     * @param data The object that is serialized and saved to a file.
      */
-    internal fun <T> save(to: FileLocation, data: T) {
-        val json = gson.toJson(data)
-        storage(to).writeText(json)
+    internal fun <T : Message> save(to: FileLocation, data: T) {
+        FileOutputStream(storage(to)).use { file ->
+            data.toByteString().writeTo(file)
+        }
     }
 
     /**
@@ -106,34 +113,6 @@ internal object FileStorage {
 }
 
 /**
- * A location of a file on disk.
- *
- * @property dir The absolute path to the directory where the file is located.
- * @property name The name of the file.
- */
-internal data class FileLocation(
-    internal val dir: String,
-    internal val name: String
-) {
-    internal companion object {
-        /**
-         * The path to application data within the user’s home directory.
-         */
-        private val appDirPath = AppDirPath.withoutVersion()
-
-        /**
-         * The location of the file with user session data.
-         */
-        internal val Session = FileLocation(appDirPath, ".session.json")
-
-        /**
-         * The location of the application settings file.
-         */
-        internal val Settings = FileLocation(appDirPath, ".settings.json")
-    }
-}
-
-/**
  * Deletes all files containing application state information
  * in the user data directory.
  *
@@ -143,24 +122,4 @@ internal data class FileLocation(
 public fun clearFileStorage() {
     FileStorage.delete(FileLocation.Session)
     FileStorage.delete(FileLocation.Settings)
-}
-
-/**
- * Provides the path to platform-specific application data
- * within the user’s home directory.
- */
-private object AppDirPath {
-    private const val author = "spine-examples"
-    private const val name = "Pingh"
-    private const val version = "1.0.0"
-
-    /**
-     * Returns the path to the application data,
-     * without including the current application version.
-     */
-    fun withoutVersion(): String {
-        val versionedPath = AppDirsFactory.getInstance()
-            .getUserDataDir(name, version, author)
-        return versionedPath.substring(0, versionedPath.length - version.length - 1)
-    }
 }
